refactor: moves cbor to eip712 converting files into eth dir
