+++ src/vscClient/eth/cborg_utils/diagnostic_test.js
