+++ src/vscClient/eth/wagmi.ts
@@ -2,7 +2,7 @@ import { encode } from '@ipld/dag-cbor'
 import { encodePayload } from 'dag-jose-utils'
 import type { Signer } from '../client'
 import { Config, signTypedData } from '@wagmi/core'
-import { convertCBORToEIP712TypedData } from '../cbor_to_eip712_converter'
+import { convertCBORToEIP712TypedData } from './cbor_to_eip712_converter'
 
 export const wagmiSigner = (async (txData, _, config: Config) => {
   const types = convertCBORToEIP712TypedData(
